document_service/requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
mongoengine==0.28.2
msgpack==1.0.8
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.2
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
